Minor changes, checking if winner and loser exists in merge method.

+++ app/models/person.rb
@@ -3,19 +3,21 @@ class Person < ActiveRecord::Base
   
   has_many :notes
 
-  def self.merge(winner, loser)  	 	
- 		result = winner
- 		result.notes << loser.notes 		
- 		result.email = loser.email if winner.email.blank?
- 		loser.destroy
- 		result
+  def self.merge(winner, loser)
+    if winner and loser  	 	
+	 		result = winner
+	 		result.notes << loser.notes 		
+	 		result.email = loser.email if winner.email.blank?
+	 		loser.destroy
+	 		result
+	 	end
   end
 
   def duplicates?(dupe)
   	if dupe
-	  	if self.first_name == dupe.first_name &&
-	  		 self.last_name == dupe.last_name &&
-	  		 self.email == dupe.email &&
+	  	if self.first_name == dupe.first_name and
+	  		 self.last_name == dupe.last_name and
+	  		 self.email == dupe.email and
 	  		 self.company_name == dupe.company_name
 	  	  return true
 	  	end
